低 - 自定义摄像机视图,还没完成

+++ OpenGLES 02/OpenGLES/Case05ViewController.m
@@ -163,7 +163,7 @@ static const SceneVertex vertices[] =
 }
 
 - (void)update{
-    float aspect = fabsf(self.view.bounds.size.width / self.view.bounds.size.height);
+    float aspect = fabs(self.view.bounds.size.width / self.view.bounds.size.height);
     
     GLKMatrix4 projectionMatrix = GLKMatrix4MakePerspective(GLKMathDegreesToRadians(65.0f), aspect, 0.1f, 100.0f);
